+++ index.html
@@ -60,7 +60,7 @@
       <div class="mt-4 bg-listBgColor rounded-md text-textDark">
         <ul class="todo-list"></ul>
         <div class="flex justify-between py-4 px-4 text-lGrayBlue">
-          <p class="items-left"></p>
+          <p class="items-left">0 items left</p>
           <button type="button" id="clear-completed-btn">
             Clear Completed
           </button>
@@ -78,7 +78,6 @@
       </ul>
     </div>
 
-    <script src="index.js"></script>
-    <script src="drag-list.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
